System.ArgumentException: messageID must have a valid value    at StreamingClient.Base.Util.Validator.ValidateString(String value, String name)    at Twitch.Base.Clients.ChatClient.<DeleteMessage>d__62.MoveNext() --- End of stack trace from previous location where exception was thrown ---    at System.Runtime.CompilerServices.TaskAwaiter.ThrowForNonSuccess(Task task)    at System.Runtime.CompilerServices.TaskAwaiter.HandleNonSuccessAndDebuggerNotification(Task task)    at MixItUp.Base.Services.Twitch.TwitchChatService.<>c__DisplayClass43_0.<<DeleteMessage>b__0>d.MoveNext() --- End of stack trace from previous location where exception was thrown ---    at System.Runtime.CompilerServices.TaskAwaiter.ThrowForNonSuccess(Task task)    at System.Runtime.CompilerServices.TaskAwaiter.HandleNonSuccessAndDebuggerNotification(Task task)    at MixItUp.Base.Util.AsyncRunner.<RunAsync>d__2.MoveNext() Full Stack:   at System.Environment.GetStackTrace(Exception e, Boolean needFileInfo)    at System.Environment.get_StackTrace()    at StreamingClient.Base.Util.Logger.Log(LogLevel level, Exception ex, Boolean includeStackTrace)    at MixItUp.Base.Util.AsyncRunner.<RunAsync>d__2.MoveNext()    at System.Runtime.CompilerServices.AsyncTaskMethodBuilder.Start[TStateMachine](TStateMachine& stateMachine)    at MixItUp.Base.Util.AsyncRunner.RunAsync(Func`1 task)    at MixItUp.Base.Services.StreamingPlatformServiceBase.<RunAsync>d__5.MoveNext()    at System.Runtime.CompilerServices.AsyncTaskMethodBuilder.Start[TStateMachine](TStateMachine& stateMachine)    at MixItUp.Base.Services.StreamingPlatformServiceBase.RunAsync(Func`1 task)    at MixItUp.Base.Services.Twitch.TwitchChatService.<DeleteMessage>d__43.MoveNext()    at System.Runtime.CompilerServices.AsyncTaskMethodBuilder.Start[TStateMachine](TStateMachine& stateMachine)    at MixItUp.Base.Services.Twitch.TwitchChatService.DeleteMessage(ChatMessageViewModel message)    at MixItUp.Base.Services.ChatService.<DeleteMessage>d__46.MoveNext()    at System.Runtime.CompilerServices.AsyncTaskMethodBuilder.Start[TStateMachine](TStateMachine& stateMachine)    at MixItUp.Base.Services.ChatService.DeleteMessage(ChatMessageViewModel message)    at MixItUp.Base.Services.ChatService.<AddMessage>d__55.MoveNext()    at System.Runtime.CompilerServices.AsyncTaskMethodBuilder.Start[TStateMachine](TStateMachine& stateMachine)    at MixItUp.Base.Services.ChatService.AddMessage(ChatMessageViewModel message)    at MixItUp.Base.Services.ChatService.<SendMessage>d__43.MoveNext()    at System.Runtime.CompilerServices.AsyncMethodBuilderCore.MoveNextRunner.InvokeMoveNext(Object stateMachine)    at System.Threading.ExecutionContext.RunInternal(ExecutionContext executionContext, ContextCallback callback, Object state, Boolean preserveSyncCtx)    at System.Threading.ExecutionContext.Run(ExecutionContext executionContext, ContextCallback callback, Object state, Boolean preserveSyncCtx)    at System.Runtime.CompilerServices.AsyncMethodBuilderCore.MoveNextRunner.Run()    at System.Threading.Tasks.SynchronizationContextAwaitTaskContinuation.<>c.<.cctor>b__8_0(Object state)    at System.Windows.Threading.ExceptionWrapper.InternalRealCall(Delegate callback, Object args, Int32 numArgs)    at System.Windows.Threading.ExceptionWrapper.TryCatchWhen(Object source, Delegate callback, Object args, Int32 numArgs, Delegate catchHandler)    at System.Windows.Threading.DispatcherOperation.InvokeImpl()    at System.Windows.Threading.DispatcherOperation.InvokeInSecurityContext(Object state)    at MS.Internal.CulturePreservingExecutionContext.CallbackWrapper(Object obj)    at System.Threading.ExecutionContext.RunInternal(ExecutionContext executionContext, ContextCallback callback, Object state, Boolean preserveSyncCtx)    at System.Threading.ExecutionContext.Run(ExecutionContext executionContext, ContextCallback callback, Object state, Boolean preserveSyncCtx)    at System.Threading.ExecutionContext.Run(ExecutionContext executionContext, ContextCallback callback, Object state)    at MS.Internal.CulturePreservingExecutionContext.Run(CulturePreservingExecutionContext executionContext, ContextCallback callback, Object state)    at System.Windows.Threading.DispatcherOperation.Invoke()    at System.Windows.Threading.Dispatcher.ProcessQueue()    at System.Windows.Threading.Dispatcher.WndProcHook(IntPtr hwnd, Int32 msg, IntPtr wParam, IntPtr lParam, Boolean& handled)    at MS.Win32.HwndWrapper.WndProc(IntPtr hwnd, Int32 msg, IntPtr wParam, IntPtr lParam, Boolean& handled)    at MS.Win32.HwndSubclass.DispatcherCallbackOperation(Object o)    at System.Windows.Threading.ExceptionWrapper.InternalRealCall(Delegate callback, Object args, Int32 numArgs)    at System.Windows.Threading.ExceptionWrapper.TryCatchWhen(Object source, Delegate callback, Object args, Int32 numArgs, Delegate catchHandler)    at System.Windows.Threading.Dispatcher.LegacyInvokeImpl(DispatcherPriority priority, TimeSpan timeout, Delegate method, Object args, Int32 numArgs)    at MS.Win32.HwndSubclass.SubclassWndProc(IntPtr hwnd, Int32 msg, IntPtr wParam, IntPtr lParam)    at MS.Win32.UnsafeNativeMethods.DispatchMessage(MSG& msg)    at System.Windows.Threading.Dispatcher.PushFrameImpl(DispatcherFrame frame)    at System.Windows.Threading.Dispatcher.PushFrame(DispatcherFrame frame)    at System.Windows.Application.RunDispatcher(Object ignore)    at System.Windows.Application.RunInternal(Window window)    at System.Windows.Application.Run(Window window)    at MixItUp.WPF.App.Main()

+++ MixItUp.Base/Services/ChatService.cs
@@ -186,7 +186,10 @@ namespace MixItUp.Base.Services
         {
             if (message.Platform == StreamingPlatformTypeEnum.Twitch)
             {
-                await this.TwitchChatService.DeleteMessage(message);
+                if (!string.IsNullOrEmpty(message.ID))
+                {
+                    await this.TwitchChatService.DeleteMessage(message);
+                }
                 if (!message.IsDeleted)
                 {
                     await message.Delete();
